Move NetBSD platform classes to oshi-common; wire NetBSD into FFM SystemInfo
NetBSD's implementation has no native dependencies (only command-line
parsing and file reads), so the classes belong next to the other "no
JNA / no FFM" platform code in oshi-common rather than oshi-core.
Mirrors the abstract-base/POJO split convention already used by
oshi-common's freebsd, openbsd, solaris, and aix packages.

- All 25 NetBSD classes (8 software/os, 13 hardware, 2 util, 2
  package-info) move from oshi-core/.../netbsd/ to
  oshi-common/.../common/.../netbsd/. The classes themselves stay
  concrete with no JNA or FFM subclass — for the now-supported FFM
  path, oshi-core-ffm SystemInfo instantiates the same shared classes.
- FFM SystemInfo gains a NETBSD case in createOperatingSystem() and
  createHardware(), and NETBSD is added to SUPPORTED_PLATFORMS. JDK 25
  on NetBSD is currently aspirational (no community port yet), but
  when one lands FFM mode will work without code changes. Issues
  #3347 (restore JNA-with-fallback for NetBSD) and #3348 (FFM-native
  bindings for NetBSD) remain open as future performance improvements.
- Replaces the CupsPrinterJNA call in NetBsdHardwareAbstractionLayer
  with a new LpstatPrinter (oshi-common) concrete CupsPrinter subclass
  that uses lpstat shell output only. NetBSD's printer enumeration is
  unchanged in behavior, just no longer coupled to oshi-core.
- FstatUtilTest switches from SystemInfo to ManagementFactory for the
  live-NetBSD PID, mirroring the FreeBSD ProcstatUtilTest pattern, so
  the test can live in oshi-common without depending on oshi-core.
- Docs (README matrix, FAQ FFM platform list, FFM module-info,
  JNA-SystemInfo javadoc) updated to reflect NetBSD in the FFM
  supported set and AIX in the FFM list it had been missing.

Tested with \`./mvnw clean install\` from root (all 788 tests pass).
