+++ oshi-common/src/main/java/oshi/hardware/common/platform/unix/netbsd/NetBsdPowerSource.java
@@ -2,7 +2,7 @@
  * Copyright 2021-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.hardware.platform.unix.netbsd;
+package oshi.hardware.common.platform.unix.netbsd;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
