+++ oshi-common/src/test/java/module-info.test
@@ -59,6 +59,12 @@
 --add-opens
   com.github.oshi.common/oshi.util.common.platform.unix.freebsd=org.junit.platform.commons
 --add-opens
+  com.github.oshi.common/oshi.hardware.common.platform.unix.netbsd=org.junit.platform.commons
+--add-opens
+  com.github.oshi.common/oshi.software.common.os.unix.netbsd=org.junit.platform.commons
+--add-opens
+  com.github.oshi.common/oshi.util.common.platform.unix.netbsd=org.junit.platform.commons
+--add-opens
   com.github.oshi.common/oshi.hardware.common.platform.unix.openbsd=org.junit.platform.commons
 --add-opens
   com.github.oshi.common/oshi.software.common.os.unix.openbsd=org.junit.platform.commons
