+++ oshi-common/src/test/java/oshi/util/common/platform/unix/netbsd/FstatUtilTest.java
@@ -2,7 +2,7 @@
  * Copyright 2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.util.platform.unix.netbsd;
+package oshi.util.common.platform.unix.netbsd;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.emptyString;
@@ -14,10 +14,12 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
+import java.lang.management.ManagementFactory;
+
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
-import oshi.SystemInfo;
+import oshi.util.ParseUtil;
 
 /**
  * Tests for {@link FstatUtil}. Pure parsing methods are covered with sample {@code fstat} output; the integration test
@@ -66,7 +68,9 @@ class FstatUtilTest {
     @Test
     @EnabledIfSystemProperty(named = "os.name", matches = "(?i)netbsd")
     void testFstatLive() {
-        int pid = new SystemInfo().getOperatingSystem().getProcessId();
+        // RuntimeMXBean#getName() returns "<pid>@<host>" — use it to avoid pulling oshi-core's SystemInfo into a
+        // oshi-common test. Mirrors the pattern used in the FreeBSD ProcstatUtilTest.
+        int pid = ParseUtil.parseIntOrDefault(ManagementFactory.getRuntimeMXBean().getName().split("@", 2)[0], -1);
         assertThat("Number of open files must be nonnegative", FstatUtil.getOpenFiles(pid),
                 is(greaterThanOrEqualTo(0L)));
         assertThat("Cwd should not be empty", FstatUtil.getCwd(pid), is(not(emptyString())));
