+++ oshi-core-ffm/src/main/java/module-info.java
@@ -2,8 +2,8 @@
  * FFM-based native implementation of the OSHI API for JDK 25+.
  * <p>
  * This module uses the Foreign Function and Memory (FFM) API for native access and currently supports Windows, macOS,
- * Linux, FreeBSD, OpenBSD, and Solaris (illumos). For broader platform support (DragonFly BSD, NetBSD, AIX), use the
- * JNA-based {@code com.github.oshi} module.
+ * Linux, FreeBSD, NetBSD, OpenBSD, Solaris (illumos), and AIX. For DragonFly BSD support, use the JNA-based
+ * {@code com.github.oshi} module.
  * <p>
  * Usage:
  *
