+++ oshi-core-ffm/src/main/java/oshi/ffm/SystemInfo.java
@@ -12,6 +12,7 @@ import java.util.function.Supplier;
 
 import oshi.annotation.PublicApi;
 import oshi.hardware.HardwareAbstractionLayer;
+import oshi.hardware.common.platform.unix.netbsd.NetBsdHardwareAbstractionLayer;
 import oshi.hardware.platform.linux.LinuxHardwareAbstractionLayerFFM;
 import oshi.hardware.platform.mac.MacHardwareAbstractionLayerFFM;
 import oshi.hardware.platform.unix.aix.AixHardwareAbstractionLayerFFM;
@@ -19,6 +20,7 @@ import oshi.hardware.platform.unix.freebsd.FreeBsdHardwareAbstractionLayerFFM;
 import oshi.hardware.platform.unix.openbsd.OpenBsdHardwareAbstractionLayerFFM;
 import oshi.hardware.platform.unix.solaris.SolarisHardwareAbstractionLayerFFM;
 import oshi.hardware.platform.windows.WindowsHardwareAbstractionLayerFFM;
+import oshi.software.common.os.unix.netbsd.NetBsdOperatingSystem;
 import oshi.software.os.OperatingSystem;
 import oshi.software.os.linux.LinuxOperatingSystemFFM;
 import oshi.software.os.mac.MacOperatingSystemFFM;
@@ -56,9 +58,9 @@ import oshi.util.PlatformEnum;
  * instance. To conserve memory at the cost of additional processing time, create a new SystemInfo for subsequent calls.
  * To conserve processing time at the cost of additional memory usage, re-use the same instance.
  * <p>
- * This implementation requires JDK 25+ and currently supports Windows, macOS, Linux, FreeBSD, OpenBSD, Solaris
+ * This implementation requires JDK 25+ and currently supports Windows, macOS, Linux, FreeBSD, NetBSD, OpenBSD, Solaris
  * (illumos), and AIX. It uses the FFM API in place of JNA for native access, which may offer better performance. For
- * broader platform support (DragonFly BSD, NetBSD), use the JNA-based entry point ({@code oshi.SystemInfo}) in the
+ * platforms not yet covered by FFM (DragonFly BSD), use the JNA-based entry point ({@code oshi.SystemInfo}) in the
  * {@code oshi-core} module.
  * <p>
  * Both this class and the JNA entry point require native access. Starting with
@@ -91,7 +93,8 @@ public class SystemInfo implements SystemInfoProvider {
     }
 
     private static final Set<PlatformEnum> SUPPORTED_PLATFORMS = EnumSet.of(PlatformEnum.LINUX, PlatformEnum.MACOS,
-            PlatformEnum.WINDOWS, PlatformEnum.FREEBSD, PlatformEnum.OPENBSD, PlatformEnum.SOLARIS, PlatformEnum.AIX);
+            PlatformEnum.WINDOWS, PlatformEnum.FREEBSD, PlatformEnum.NETBSD, PlatformEnum.OPENBSD, PlatformEnum.SOLARIS,
+            PlatformEnum.AIX);
 
     @Override
     public boolean isAvailable() {
@@ -108,6 +111,8 @@ public class SystemInfo implements SystemInfoProvider {
                 return new WindowsOperatingSystemFFM();
             case FREEBSD:
                 return new FreeBsdOperatingSystemFFM();
+            case NETBSD:
+                return new NetBsdOperatingSystem();
             case OPENBSD:
                 return new OpenBsdOperatingSystemFFM();
             case SOLARIS:
@@ -129,6 +134,8 @@ public class SystemInfo implements SystemInfoProvider {
                 return new WindowsHardwareAbstractionLayerFFM();
             case FREEBSD:
                 return new FreeBsdHardwareAbstractionLayerFFM();
+            case NETBSD:
+                return new NetBsdHardwareAbstractionLayer();
             case OPENBSD:
                 return new OpenBsdHardwareAbstractionLayerFFM();
             case SOLARIS:
